[#24532] Stiffness recovery for the ENDO_FISS_EXP damage law
The ENDO_FISS_EXP damage law now models the stiffness recovery observed when
closing micro-cracks through compressive strain states. In order to smooth the
transition between damaged and recovered stiffnesses, a regularisation
function is introduced which depends on a single parameter characterising the
strength of the regularisation. It is denoted REST_RIGIDITE in the command
DEFI_MATERIAU (key-word ENDO_FISS_EXP) and it is mandatory. This parameter can
also be set with the macro-command DEFI_MATER_GC through the key-word
REST_RIGI_FC which is optional with default value 0.9. In this case, one
indicates how many stiffness is recovered relatively to full recovery at a
strain state equal to fc/E where fc is the compressive strength and E the
Young modulus.

Moreover, the macro-command DEFI_MATER_GC has been slightly alterned. Indeed,
because of the stiffness recovery, the consistence of the damage law with a
cohesive law is now established in the case of a confined uniaxial tension
test (and no more a free uniaxial tension test). Therefore, the computation of
the internal parameters of the model from the macroscopic ones provided in
DEFI_MATER_GC is no more exactly the same: for the same set of parameters in
DEFI_MATER_GC, the results are now slightly different (and more than slightly
is stiffness recovery occurs, of course).

Besides, it was the opportunity to modify the keywords in DEFI_MATER_GC. Now,
the keyword FT_FENDAGE is removed (because confusing and useless) and DSIG_DU
is replaced by G_INIT which provides a more classical way to describe the
initial slope of the cohesive law through the resulting fracture energy that
would be obtained with a linear softening law.

Finally, two errors have been corrected, one when computing the tangent
stiffness, another in the path-following method PRED_ELAS related to the
ENDO_FISS_EXP law. None of them led to erroneous results but they might hinder
the convergence of the solution algorithm. In addition, the secant stiffness
matrix has been introduced. It required the extension of the internal
variable set in order to memorize the previous strain state when computing
a pseudo-secant for the stiffness recovery regularisation function.

The developments have been widely tested during the validation phase of the
model. In terms of verification, the same tests as previously are used, with
an adjustment of the non-regression values because of the alteration of
DEFI_MATER_GC. In addition, the closed-form solution in SSNV234 is revisited
for the same reason; moreover, it includes now a compressive phase in order
to test the stiffness recovery.


Former-commit-id: 52cfc01a8158553aa48b83458b42ee3552a12cfa