[#24829] Macro CALC_ENDO
A new macro-command CALC_ENDO is introduced in bibpyt/Contrib which:
1) Reads a mesh and a damage field (possibly void) in the same MED file
2) Finds an equilibrium with the damage field
3) Propagates damage along a sequence of loading with a path_following method
4) Export the results (displacements, stresses and damage)
The macro-command is focused on robustness with different strategies to ensure
the convergence. It is restricted to the ENDO_FISS_EXP damage law.

In addition, it is now possible to compute the strain field in post-processing
of a gradient damage computation:
- Infinitesimal strain located at the Gauss points or the nodes
- Green-Lagrange strain located at the Gauss points or the nodes
- strain invariants located at the Gauss points or the nodes


Former-commit-id: cb3bff19457c5bbab349f57464a89dc5b9ba514a